fix: refactor commitments package (#726)
