+++ internal/scheduling/reservations/commitments/api/info_monitor.go
@@ -1,7 +1,7 @@
 // Copyright SAP SE
 // SPDX-License-Identifier: Apache-2.0
 
-package commitments
+package api
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
