+++ internal/scheduling/reservations/commitments/api/info_test.go
@@ -1,7 +1,7 @@
 // Copyright SAP SE
 // SPDX-License-Identifier: Apache-2.0
 
-package commitments
+package api
 
 import (
 	"encoding/json"
