+++ internal/scheduling/reservations/commitments/state.go
@@ -46,10 +46,10 @@ func ResourceNameInstances(flavorGroup string) string {
 	return resourceNamePrefix + flavorGroup + ResourceSuffixInstances
 }
 
-// getFlavorGroupNameFromResource extracts the flavor group name from a LIQUID resource name.
+// GetFlavorGroupNameFromResource extracts the flavor group name from a LIQUID resource name.
 // Only accepts _ram resources since CommitmentState is RAM-based.
 // Callers handling _cores or _instances must use a different approach.
-func getFlavorGroupNameFromResource(resourceName string) (string, error) {
+func GetFlavorGroupNameFromResource(resourceName string) (string, error) {
 	if !strings.HasPrefix(resourceName, resourceNamePrefix) {
 		return "", fmt.Errorf("invalid resource name: %s (missing prefix)", resourceName)
 	}
@@ -105,7 +105,7 @@ func FromCommitment(
 		return nil, errors.New("unexpected commitment format")
 	}
 
-	flavorGroupName, err := getFlavorGroupNameFromResource(commitment.ResourceName)
+	flavorGroupName, err := GetFlavorGroupNameFromResource(commitment.ResourceName)
 	if err != nil {
 		return nil, err
 	}
@@ -205,41 +205,6 @@ func FromChangeCommitmentTargetState(
 		StartTime:        startTime,
 		EndTime:          endTime,
 	}, nil
-}
-
-// CommitmentStateWithUsage extends CommitmentState with usage tracking for billing calculations.
-// Used by the report-usage API to track remaining capacity during VM-to-commitment assignment.
-type CommitmentStateWithUsage struct {
-	CommitmentState
-	// RemainingMemoryBytes is the uncommitted capacity left for VM assignment
-	RemainingMemoryBytes int64
-	// AssignedVMs tracks which VMs have been assigned to this commitment
-	AssignedVMs []string
-}
-
-// NewCommitmentStateWithUsage creates a CommitmentStateWithUsage from a CommitmentState.
-func NewCommitmentStateWithUsage(state *CommitmentState) *CommitmentStateWithUsage {
-	return &CommitmentStateWithUsage{
-		CommitmentState:      *state,
-		RemainingMemoryBytes: state.TotalMemoryBytes,
-		AssignedVMs:          []string{},
-	}
-}
-
-// AssignVM attempts to assign a VM to this commitment if there's enough capacity.
-// Returns true if the VM was assigned, false if not enough capacity.
-func (c *CommitmentStateWithUsage) AssignVM(vmUUID string, vmMemoryBytes int64) bool {
-	if c.RemainingMemoryBytes >= vmMemoryBytes {
-		c.RemainingMemoryBytes -= vmMemoryBytes
-		c.AssignedVMs = append(c.AssignedVMs, vmUUID)
-		return true
-	}
-	return false
-}
-
-// HasRemainingCapacity returns true if the commitment has any remaining capacity.
-func (c *CommitmentStateWithUsage) HasRemainingCapacity() bool {
-	return c.RemainingMemoryBytes > 0
 }
 
 // FromReservations reconstructs CommitmentState from existing Reservation CRDs.
